elastic_wifi/requirements.txt
chalice==1.3.0
boto==2.48.0
Flask==1.0.2
requests==2.19.1
urllib3==1.23
geopy==1.14.0
elasticsearch==6.3.0
